Issue (#355) Move Mavis into PegasusState
DO NOT MERGE THIS!!!!!!!!!!! Needs MAVIS update
