+++ core/PegasusState.cpp
@@ -72,6 +72,12 @@ namespace pegasus
         validate_inst_begin_(p->validate_inst_begin),
         validate_fail_on_first_diff_(p->validate_fail_on_first_diff),
         priv_mode_(getPrivilegeMode(p->priv_mode)),
+        isa_string_(hart_tn->getParent()->getChildAs<sparta::ParameterSet>("params")->getParameterValueAs<std::string>("isa")),
+        isa_file_path_(hart_tn->getParent()->getChildAs<sparta::ParameterSet>("params")->getParameterValueAs<std::string>("isa_file_path")),
+        arch_name_(hart_tn->getParent()->getChildAs<sparta::ParameterSet>("params")->getParameterValueAs<std::string>("arch")),
+        uarch_file_path_(hart_tn->getParent()->getChildAs<sparta::ParameterSet>("params")->getParameterValueAs<std::string>("uarch_file_path")),
+        extension_manager_(mavis::extension_manager::riscv::RISCVExtensionManager::fromISA(isa_string_,
+            isa_file_path_ + std::string("/riscv_isa_spec.json"), isa_file_path_)),
         inst_logger_(hart_tn, "inst", "Pegasus Instruction Logger"),
         stf_valid_logger_(hart_tn, "stf_valid", "Pegasus STF Validator Logger"),
         finish_action_group_("finish_inst"),
@@ -120,6 +126,17 @@ namespace pegasus
             sparta_assert(false, "Unsupported XLEN");
         }
 
+        if (isCompressionEnabled())
+        {
+            setPcAlignment_(2);
+        }
+        else
+        {
+            setPcAlignment_(4);
+        }
+
+        init_csr_enabled_state_();
+
         // Increment PC Action
         const bool CHECK_ILIMIT = ilimit_ > 0;
         if (CHECK_ILIMIT)
@@ -150,6 +167,27 @@ namespace pegasus
         vector_config_.setVLEN(vlen_);
     }
 
+    mavis::FileNameListType PegasusState::getUArchFiles_() const
+    {
+        mavis::FileNameListType uarch_files;
+
+        const std::string xlen_str = std::to_string(xlen_);
+        const std::string xlen_uarch_file_path =
+            uarch_file_path_ + "/" + arch_name_ + "/rv" + xlen_str + "/gen";
+        const std::regex filename_regex("pegasus_uarch_.*json");
+        for (const auto & entry : std::filesystem::directory_iterator{xlen_uarch_file_path})
+        {
+            if (std::regex_search(entry.path().filename().string(), filename_regex))
+            {
+                DLOG("Loading: " << entry.path());
+                uarch_files.emplace_back(entry.path());
+            }
+        }
+
+        return uarch_files;
+    }
+
+
     // Not default -- defined in source file to reduce massive inlining
     PegasusState::~PegasusState() {}
 
@@ -166,6 +204,25 @@ namespace pegasus
 
         // Connect finish ActionGroup to Fetch
         finish_action_group_.setNextActionGroup(fetch_unit_->getActionGroup());
+
+        // Initialize Mavis
+        DLOG("Initializing Mavis with ISA string " << isa_string_);
+
+        mavis_ = std::make_unique<MavisType>(
+            extension_manager_.constructMavis<
+                PegasusInst, PegasusExtractor, PegasusInstAllocatorWrapper<PegasusInstAllocator>,
+                PegasusExtractorAllocatorWrapper<PegasusExtractorAllocator>>(
+                getUArchFiles_(), mavis_uid_list_, {}, // annotation overrides
+                {},                                    // inclusions
+                {},                                    // exclusions
+                PegasusInstAllocatorWrapper<PegasusInstAllocator>(
+                    sparta::notNull(PegasusAllocators::getAllocators(getContainer()))
+                        ->inst_allocator),
+                PegasusExtractorAllocatorWrapper<PegasusExtractorAllocator>(
+                    sparta::notNull(PegasusAllocators::getAllocators(getContainer()))
+                        ->extractor_allocator,
+                    this)));
+
     }
 
     void PegasusState::onBindTreeLate_()
@@ -364,6 +421,20 @@ namespace pegasus
         }
 
         return ++action_it;
+    }
+
+    void PegasusState::changeMavisContext()
+    {
+        extension_manager_.switchMavisContext(*mavis_.get());
+
+        if (isCompressionEnabled())
+        {
+            setPcAlignment_(2);
+        }
+        else
+        {
+            setPcAlignment_(4);
+        }
     }
 
     void PegasusState::enableInteractiveMode()
@@ -824,19 +895,19 @@ namespace pegasus
         const auto COMPAT_HWCAP_ISA_V(1 << ('V' - 'A'));
 
         uint64_t a_val = COMPAT_HWCAP_ISA_I | COMPAT_HWCAP_ISA_M | COMPAT_HWCAP_ISA_A;
-        if (pegasus_core_->isExtensionEnabled("f"))
+        if (isExtensionEnabled("f"))
         {
             a_val |= COMPAT_HWCAP_ISA_F;
         }
-        if (pegasus_core_->isExtensionEnabled("d"))
+        if (isExtensionEnabled("d"))
         {
             a_val |= COMPAT_HWCAP_ISA_D;
         }
-        if (pegasus_core_->isExtensionEnabled("c"))
+        if (isExtensionEnabled("c"))
         {
             a_val |= COMPAT_HWCAP_ISA_C;
         }
-        if (pegasus_core_->isExtensionEnabled("v"))
+        if (isExtensionEnabled("v"))
         {
             a_val |= COMPAT_HWCAP_ISA_V;
         }
@@ -978,9 +1049,9 @@ namespace pegasus
     template bool PegasusState::SimState::compare<false>(const SimState* rhs) const;
     template bool PegasusState::SimState::compare<true>(const SimState* rhs) const;
 
-    void PegasusState::init_csr_enabled_state()
+    void PegasusState::init_csr_enabled_state_()
     {
-        const auto & extensionManager = getCore()->getExtensionManager();
+        const auto & extensionManager = getExtensionManager();
 
         // Enable all register by default
         csr_enabled_state_.resize(csr_rset_->size(), true);
