+++ test/sim/InstructionTester.hpp
@@ -77,7 +77,7 @@ class PegasusInstructionTester
         pegasus::PegasusState* state = getPegasusState();
         mavis::ExtractorDirectInfo ex_info(args...);
         pegasus::PegasusInst::PtrType instPtr =
-            state->getCore()->getMavis()->makeInstDirectly(ex_info, state);
+            state->getMavis()->makeInstDirectly(ex_info, state);
         std::cout << instPtr->dasmString() << std::endl;
         return instPtr;
     }
